fix: URLImageFeatureをファイルに分割
